BaiTap6Chieu2

+++ src/main/java/com/example/bai5/repository/CategoryRepository.java
@@ -1,7 +1,8 @@
 package com.example.bai5.repository;
 
-import com.example.bai5.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+import com.example.bai5.model.Category;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
